backend/python_service/requirements.txt
flask
ultralytics
opencv-python
numpy
pillow
pyyaml
